[b-10] Add, Edit movie:  - MovieAdminController has been added to serve the purpose of adding and editing movies;  - this controller serves requests only from admin users;  - EnableGlobalMethodSecurity option is added to Security configuration to protect MovieAdminController methods;  - ChangeMovieDto class added that represents request from user for editing or adding movie;  - Genre and Country classes are marked as Hibernate entity;  - Parallel enrichment service implementations were changed in accordance to hibernate DAO implementations;

+++ src/main/java/edu/ted/web/movieland/dao/MovieDao.java
@@ -10,7 +10,7 @@ public interface MovieDao {
 
     List<Movie> findAll(Sorting sorting);
     List<Movie> getNRandomMovies(int count);
-    List<Movie> getMoviesByGenre(int genreId, Sorting sorting);
+    List<Movie> getMoviesByGenre(long genreId, Sorting sorting);
     Optional<Movie> getMovieById(long movieId);
     Movie saveOrUpdate(Movie movie);
 }
